Spinner, using HOC in TimeseriesChartMeta
