+++ src/components/TimeseriesChartMeta/TimeseriesChartMetaPure.test.tsx
@@ -3,7 +3,7 @@ import { configure, mount, shallow } from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
 import React from 'react';
 import { datapoints, timeseriesList } from '../../mocks';
-import { TimeseriesChartMeta } from './TimeseriesChartMeta';
+import { TimeseriesChartMetaPure } from './TimeseriesChartMetaPure';
 
 configure({ adapter: new Adapter() });
 
@@ -26,7 +26,9 @@ describe('TimeseriesChartMeta', () => {
   const activeLabelSelector = 'label.ant-radio-button-wrapper-checked';
 
   it('Should render without exploding', () => {
-    const wrapper = shallow(<TimeseriesChartMeta timeseries={timeseries} />);
+    const wrapper = shallow(
+      <TimeseriesChartMetaPure timeseries={timeseries} />
+    );
     expect(wrapper.find('RadioGroup')).toHaveLength(1);
     expect(wrapper.find('TimeseriesChart')).toHaveLength(1);
     expect(wrapper.find('TimeseriesValue')).toHaveLength(1);
@@ -35,7 +37,7 @@ describe('TimeseriesChartMeta', () => {
 
   it('Should not render elements if they are hidden', () => {
     const wrapper = shallow(
-      <TimeseriesChartMeta
+      <TimeseriesChartMetaPure
         showPeriods={false}
         showChart={false}
         showDatapoint={false}
@@ -50,7 +52,7 @@ describe('TimeseriesChartMeta', () => {
   });
 
   it('Should have default period 1 hour', () => {
-    const wrapper = mount(<TimeseriesChartMeta timeseries={timeseries} />);
+    const wrapper = mount(<TimeseriesChartMetaPure timeseries={timeseries} />);
     expect(wrapper).toHaveLength(1);
     const checkedLabel = wrapper.find(activeLabelSelector);
     expect(checkedLabel).toHaveLength(1);
@@ -59,7 +61,7 @@ describe('TimeseriesChartMeta', () => {
 
   it('Should not have active period selected is defaultBasePeriod has been provided', () => {
     const wrapper = mount(
-      <TimeseriesChartMeta
+      <TimeseriesChartMetaPure
         timeseries={timeseries}
         defaultBasePeriod={{
           startTime: Date.now() - 1000000,
@@ -72,7 +74,7 @@ describe('TimeseriesChartMeta', () => {
   });
 
   it('Should switch period on click', () => {
-    const wrapper = mount(<TimeseriesChartMeta timeseries={timeseries} />);
+    const wrapper = mount(<TimeseriesChartMetaPure timeseries={timeseries} />);
     const radioInputs = wrapper.find('input.ant-radio-button-input');
     radioInputs.first().simulate('change', { target: { checked: true } });
     const checkedLabel = wrapper.find(activeLabelSelector);
@@ -82,7 +84,7 @@ describe('TimeseriesChartMeta', () => {
 
   it('Should render nothing if timeseries is null or undefined', () => {
     // @ts-ignore
-    const wrapper = mount(<TimeseriesChartMeta timeseries={null} />);
+    const wrapper = mount(<TimeseriesChartMetaPure timeseries={null} />);
     expect(wrapper.isEmptyRender()).toBeTruthy();
   });
 });
