+++ src/components/TimeseriesChartMeta/stories/TimeseriesChartMeta.stories.tsx
@@ -1,7 +1,8 @@
 /* eslint-disable react/no-multi-comp */
+import * as sdk from '@cognite/sdk';
 import { storiesOf } from '@storybook/react';
 import React from 'react';
-import { setupMocks } from '../../TimeseriesChart/stories/TimeseriesChart.stories';
+import { setupMocks as setupTimeseriesChartMocks } from '../../TimeseriesChart/stories/TimeseriesChart.stories';
 import { TimeseriesChartMeta } from '../TimeseriesChartMeta';
 
 import customBasePeriod from './customBasePeriod.md';
@@ -11,28 +12,40 @@ import full from './full.md';
 import hideElements from './hideElements.md';
 import predefinedPeriod from './predefinedPeriod.md';
 
-const ts = {
-  id: 8681821313339919,
-  name: 'IA_21PT1019.AlarmByte',
-  isString: false,
-  unit: 'bar',
-  metadata: {
-    tag: 'IA_21PT1019.AlarmByte',
-    scan: '1',
-    span: '100',
-    step: '1',
-    zero: '0',
-  },
-  assetId: 4965555138606429,
-  isStep: false,
-  description: '21PT1019.AlarmByte',
+const setupMocks = () => {
+  setupTimeseriesChartMocks();
+  // @ts-ignore
+  sdk.TimeSeries.retrieve = () => {
+    return new Promise(resolve => {
+      setTimeout(
+        () =>
+          resolve({
+            id: 8681821313339919,
+            name: 'IA_21PT1019.AlarmByte',
+            isString: false,
+            unit: 'bar',
+            metadata: {
+              tag: 'IA_21PT1019.AlarmByte',
+              scan: '1',
+              span: '100',
+              step: '1',
+              zero: '0',
+            },
+            assetId: 4965555138606429,
+            isStep: false,
+            description: '21PT1019.AlarmByte',
+          }),
+        1000
+      );
+    });
+  };
 };
 
 storiesOf('TimeseriesChartMeta', module).add(
   'Full description',
   () => {
     setupMocks();
-    return <TimeseriesChartMeta timeseries={ts} />;
+    return <TimeseriesChartMeta timeseriesId={123} />;
   },
   {
     readme: {
@@ -48,7 +61,7 @@ storiesOf('TimeseriesChartMeta/Examples', module)
     () => {
       setupMocks();
       return (
-        <TimeseriesChartMeta timeseries={ts} defaultTimePeriod="lastMonth" />
+        <TimeseriesChartMeta timeseriesId={123} defaultTimePeriod="lastMonth" />
       );
     },
     {
@@ -63,7 +76,7 @@ storiesOf('TimeseriesChartMeta/Examples', module)
       setupMocks();
       return (
         <TimeseriesChartMeta
-          timeseries={ts}
+          timeseriesId={123}
           showChart={true}
           showDescription={false}
           showDatapoint={false}
@@ -82,7 +95,7 @@ storiesOf('TimeseriesChartMeta/Examples', module)
     'Disable live updates',
     () => {
       setupMocks();
-      return <TimeseriesChartMeta timeseries={ts} liveUpdate={false} />;
+      return <TimeseriesChartMeta timeseriesId={123} liveUpdate={false} />;
     },
     {
       readme: {
@@ -96,7 +109,7 @@ storiesOf('TimeseriesChartMeta/Examples', module)
       setupMocks();
       return (
         <TimeseriesChartMeta
-          timeseries={ts}
+          timeseriesId={123}
           liveUpdate={true}
           updateIntervalMillis={1000}
         />
@@ -114,7 +127,7 @@ storiesOf('TimeseriesChartMeta/Examples', module)
       setupMocks();
       return (
         <TimeseriesChartMeta
-          timeseries={ts}
+          timeseriesId={123}
           defaultBasePeriod={{
             startTime: Date.now() - 10 * 24 * 60 * 60 * 1000, // 10 days ago
             endTime: Date.now(),
