+++ src/components/common/LoadingOverlay/LoadingOverlay.tsx
@@ -2,11 +2,11 @@ import { Spin } from 'antd';
 import React from 'react';
 import styled from 'styled-components';
 
-const SpinContainer = styled(Spin)`
+const SpinContainer = styled(Spin)<{ backgroundColor?: string }>`
   width: 100%;
   height: 100%;
   position: absolute !important;
-  background-color: rgba(0, 0, 0, 0.5);
+  background-color: ${({ backgroundColor }) => backgroundColor};
   z-index: 100;
 
   &:before {
@@ -17,19 +17,31 @@ const SpinContainer = styled(Spin)`
   }
 `;
 
+SpinContainer.defaultProps = {
+  backgroundColor: 'rgba(0, 0, 0, 0.5)',
+};
+
 interface LoadingOverlayProps {
   isLoading: boolean;
   delay?: number;
   size?: 'large' | 'small' | 'default';
+  backgroundColor?: string;
 }
 
 export function LoadingOverlay({
   isLoading,
+  backgroundColor,
   delay = 0,
   size = 'default',
 }: LoadingOverlayProps) {
   if (isLoading) {
-    return <SpinContainer delay={delay} size={size} />;
+    return (
+      <SpinContainer
+        delay={delay}
+        size={size}
+        backgroundColor={backgroundColor}
+      />
+    );
   }
 
   return null;
